Updated the README to reflect that this is a maintained fork of Yagops bot, and not a fork used for contributing.
If you want to contribute to Yagops source repository - as in a plugin.
Ensure you use the `upstream` branch as your source.

https://github.com/chuckbutler/telegram-bot/tree/upstream

like so:

    git fetch
    git checkout upstream
    git checkout -b my-feature
    *hack hack hack*
    git commit

from here the rest is the normal github workflow to submit a Pull
Request

The upstream branch will be responsible for backporting upstream changes
to rebase into our fork as they cut new releases.